Add explicit component `power-select-multiple-with-create`
This removes the option to pass `multiple=true` into the
`power-select-with-create` component.
