+++ addon/components/power-select-multiple-with-create.js
@@ -1,0 +1,19 @@
+import PowerSelectWithCreate from './power-select-with-create';
+import layout from '../templates/components/power-select-with-create';
+
+export default PowerSelectWithCreate.extend({
+  layout,
+  powerSelectComponentName: 'power-select-multiple',
+
+  selectOrCreate(selection, select) {
+    let suggestion = selection.filter((option) => {
+      return option.__isSuggestion__;
+    })[0];
+
+    if (suggestion) {
+      this.get('oncreate')(suggestion.__value__, select);
+    } else {
+      this.get('onchange')(selection, select);
+    }
+  },
+});
