+++ addon/templates/components/power-select-with-create.hbs
@@ -1,104 +1,52 @@
-{{#if multiple}}
-  {{#power-select-multiple
-    ariaDescribedBy=ariaDescribedBy
-    ariaInvalid=ariaInvalid
-    ariaLabel=ariaLabel
-    ariaLabelledBy=ariaLabelledBy
-    options=optionsArray
-    selected=selected
-    onchange=(action "selectOrCreate")
-    onkeydown=onkeydown
-    onfocus=onfocus
-    onopen=onopen
-    onclose=onclose
-    search=(action "searchAndSuggest")
-    disabled=disabled
-    placeholder=placeholder
-    searchEnabled=searchEnabled
-    searchPlaceholder=searchPlaceholder
-    searchField=searchField
-    matcher=matcher
-    loadingMessage=loadingMessage
-    noMatchesMessage=noMatchesMessage
-    onblur=onblur
-    searchMessage=searchMessage
-    triggerComponent=triggerComponent
-    selectedItemComponent=selectedItemComponent
-    afterOptionsComponent=afterOptionsComponent
-    beforeOptionsComponent=beforeOptionsComponent
-    optionsComponent=optionsComponent
-    renderInPlace=renderInPlace
-    closeOnSelect=closeOnSelect
-    allowClear=allowClear
-    verticalPosition=verticalPosition
-    horizontalPosition=horizontalPosition
-    class=class
-    triggerClass=triggerClass
-    dropdownClass=dropdownClass
-    extra=extra
-    initiallyOpened=initiallyOpened
-    matchTriggerWidth=matchTriggerWidth
-    destination=destination
-    tabindex=tabindex
-    dir=dir
-    defaultHighlighted=defaultHighlighted
-    as |option term|}}
-    {{#if option.__isSuggestion__}}
-      {{component suggestedOptionComponent option=option term=term}}
-    {{else}}
-      {{yield option term}}
-    {{/if}}
-  {{/power-select-multiple}}
-{{else}}
-  {{#power-select
-    ariaDescribedBy=ariaDescribedBy
-    ariaInvalid=ariaInvalid
-    ariaLabel=ariaLabel
-    ariaLabelledBy=ariaLabelledBy
-    options=optionsArray
-    selected=selected
-    onchange=(action "selectOrCreate")
-    onkeydown=onkeydown
-    oninput=oninput
-    onfocus=onfocus
-    onopen=onopen
-    onclose=onclose
-    search=(action "searchAndSuggest")
-    disabled=disabled
-    placeholder=placeholder
-    searchEnabled=searchEnabled
-    searchPlaceholder=searchPlaceholder
-    searchField=searchField
-    matcher=matcher
-    loadingMessage=loadingMessage
-    noMatchesMessage=noMatchesMessage
-    onblur=onblur
-    searchMessage=searchMessage
-    triggerComponent=triggerComponent
-    selectedItemComponent=selectedItemComponent
-    afterOptionsComponent=afterOptionsComponent
-    beforeOptionsComponent=beforeOptionsComponent
-    optionsComponent=optionsComponent
-    renderInPlace=renderInPlace
-    closeOnSelect=closeOnSelect
-    allowClear=allowClear
-    verticalPosition=verticalPosition
-    horizontalPosition=horizontalPosition
-    class=class
-    triggerClass=triggerClass
-    dropdownClass=dropdownClass
-    extra=extra
-    initiallyOpened=initiallyOpened
-    matchTriggerWidth=matchTriggerWidth
-    destination=destination
-    tabindex=tabindex
-    dir=dir
-    defaultHighlighted=defaultHighlighted
-    as |option term|}}
-    {{#if option.__isSuggestion__}}
-      {{component suggestedOptionComponent option=option term=term}}
-    {{else}}
-      {{yield option term}}
-    {{/if}}
-  {{/power-select}}
-{{/if}}
+{{#component powerSelectComponentName
+  ariaDescribedBy=ariaDescribedBy
+  ariaInvalid=ariaInvalid
+  ariaLabel=ariaLabel
+  ariaLabelledBy=ariaLabelledBy
+  options=optionsArray
+  selected=selected
+  onchange=(action selectOrCreate)
+  onkeydown=onkeydown
+  oninput=oninput
+  onfocus=onfocus
+  onopen=onopen
+  onclose=onclose
+  search=(action searchAndSuggest)
+  disabled=disabled
+  placeholder=placeholder
+  searchEnabled=searchEnabled
+  searchPlaceholder=searchPlaceholder
+  searchField=searchField
+  matcher=matcher
+  loadingMessage=loadingMessage
+  noMatchesMessage=noMatchesMessage
+  onblur=onblur
+  searchMessage=searchMessage
+  triggerComponent=triggerComponent
+  selectedItemComponent=selectedItemComponent
+  afterOptionsComponent=afterOptionsComponent
+  beforeOptionsComponent=beforeOptionsComponent
+  optionsComponent=optionsComponent
+  renderInPlace=renderInPlace
+  closeOnSelect=closeOnSelect
+  allowClear=allowClear
+  verticalPosition=verticalPosition
+  horizontalPosition=horizontalPosition
+  class=class
+  triggerClass=triggerClass
+  dropdownClass=dropdownClass
+  extra=extra
+  initiallyOpened=initiallyOpened
+  matchTriggerWidth=matchTriggerWidth
+  destination=destination
+  tabindex=tabindex
+  dir=dir
+  defaultHighlighted=defaultHighlighted
+  as |option term|
+}}
+  {{#if option.__isSuggestion__}}
+    {{component suggestedOptionComponent option=option term=term}}
+  {{else}}
+    {{yield option term}}
+  {{/if}}
+{{/component}}
