+++ app/components/power-select-multiple-with-create.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-power-select-with-create/components/power-select-multiple-with-create';
