+++ tests/dummy/app/templates/application.hbs
@@ -14,6 +14,20 @@
 <br>
 <br>
 <br>
+<h3>Multiple selections via <code>power-select-multiple-with-create</code></h3>
+{{#power-select-multiple-with-create
+    options=countries
+    searchField="name"
+    selected=selectedCountries
+    onchange=(action (mut selectedCountries))
+    oncreate=(action "createCountry") as |country term|
+}}
+  {{country.name}}
+{{/power-select-multiple-with-create}}
+
+<br>
+<br>
+<br>
 <h3>With custom Add message</h3>
 {{#power-select-with-create
     options=countries
