+++ tests/integration/components/power-select-with-create-test.js
@@ -212,117 +212,6 @@ test('async search works with an ArrayProxy', function(assert) {
   assert.equal(options.eq(2).text().trim(), 'Bar');
 });
 
-test('it displays multiple selections correctly', function(assert) {
-  assert.expect(2);
-
-  const initialSelection = [this.get('countries.firstObject')];
-  this.set('selectedCountries', initialSelection);
-
-  this.render(hbs`
-    {{#power-select-with-create
-        options=countries
-        selected=selectedCountries
-        onchange=(action (mut selectedCountries))
-        oncreate=(action "createCountry")
-        multiple=true as |country|
-    }}
-      {{country.name}}
-    {{/power-select-with-create}}
-  `);
-
-  const selectedOptions = this.$('.ember-power-select-multiple-option');
-  assert.equal(selectedOptions.length, initialSelection.length);
-  assert.ok(selectedOptions.text().indexOf(initialSelection[0].name) !== -1);
-});
-
-test('it passes an array to onchange in multiple mode', function(assert) {
-  assert.expect(4);
-
-  this.set('selectedCountries', []);
-  this.on('selectCountries', function(countries) {
-    assert.ok(countries instanceof Array);
-    this.set('selectedCountries', countries);
-  });
-
-  this.render(hbs`
-    {{#power-select-with-create
-        options=countries
-        selected=selectedCountries
-        onchange=(action "selectCountries")
-        oncreate=(action "createCountry")
-        multiple=true as |country|
-    }}
-      {{country.name}}
-    {{/power-select-with-create}}
-  `);
-
-  clickTrigger();
-  nativeMouseUp('.ember-power-select-option:eq(0)');
-
-  assert.equal(this.get('selectedCountries.length'), 1);
-
-  clickTrigger();
-  nativeMouseUp('.ember-power-select-option:eq(1)');
-
-  assert.equal(this.get('selectedCountries.length'), 2);
-});
-
-test('it calls oncreate correctly in multiple mode', function(assert) {
-  assert.expect(1);
-
-  this.set('selectedCountries', []);
-  this.on('createCountry', function(country) {
-    assert.equal(country, 'Foo Bar');
-  });
-
-  this.render(hbs`
-    {{#power-select-with-create
-        options=countries
-        selected=selectedCountries
-        oncreate=(action "createCountry")
-        multiple=true as |country|
-    }}
-      {{country.name}}
-    {{/power-select-with-create}}
-  `);
-
-  clickTrigger();
-  Ember.run(() => typeInSearch('Foo Bar'));
-  nativeMouseUp('.ember-power-select-option:eq(0)');
-});
-
-test('it supports async search function', function(assert) {
-  this.set('selectedCountries', []);
-  this.on('searchCountries', () => {
-    return new Ember.RSVP.Promise((resolve) => {
-      resolve([{name: 'Foo'}, {name: 'Bar'}]);
-    });
-  });
-
-  this.render(hbs`
-    {{#power-select-with-create
-        search=(action "searchCountries")
-        selected=selectedCountries
-        onchange=(action (mut selectedCountries))
-        oncreate=(action "createCountry")
-        multiple=true as |country|
-    }}
-      {{country.name}}
-    {{/power-select-with-create}}
-  `);
-
-  clickTrigger();
-  typeInSearch('foo');
-  nativeMouseUp('.ember-power-select-option:eq(1)');
-
-  clickTrigger();
-  typeInSearch('foo');
-  nativeMouseUp('.ember-power-select-option:eq(2)');
-
-  assert.equal(this.get('selectedCountries')[0].name, 'Foo');
-  assert.equal(this.get('selectedCountries')[1].name, 'Bar');
-});
-
 test('it lets the user decide if the create option should be shown', function(assert) {
   assert.expect(5);
